+++ packages/ui/src/components/index.ts
@@ -23,6 +23,7 @@ export {
 	IconPullRequest,
 	IconQuestion,
 	IconSettings,
+	IconShadow,
 	IconSun,
 	IconTimeline,
 	IconWarning
@@ -30,10 +31,10 @@ export {
 export { Icon, type IconProps } from "./icon/icon";
 export { InputCheckbox } from "./input-checkbox/input-checkbox";
 export { InputDatetime } from "./input-datetime/input-datetime";
-export { InputNumber } from "./input-number/input-number";
+export { InputLocation } from "./input-location/input-location";
 export { InputRadio } from "./input-radio/input-radio";
 export { InputSelect, type InputSelectOption } from "./input-select/input-select";
-export { InputText } from "./input-text/input-text";
+export { Input } from "./input/input";
 export { Link, type LinkProps } from "./link/link";
 export { Loader } from "./loader/loader";
 export { Modal } from "./modal/modal";