+++ packages/photo-companion/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.27.0
+
+- Fixed incorrect sun azimuth numeric value;
+- New section: height by shadow length calculator;
+
 ## 2.26.0
 
 - Fix modal and drawer height on mobile, panel height are taken into account;